--- frontend/src/pages/selectAreaPage/selectAreaPage.module.css
@@ -1,31 +1,0 @@
-.container {
-  margin: 0;
-  padding: 0;
-}
-
-/* Body */
-.body {
-  display: flex;
-  justify-content: center;
-}
-
-.areaList {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  list-style: none;
-}
-
-.area {
-  border: outset;
-}
-
-.area:hover {
-  background-color: skyblue;
-}
-
-/* 지역 클릭 후 */
-.areaBox {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
