+++ frontend/src/pages/selectContainer/selectArea/showAll/showAll.jsx
@@ -1,0 +1,7 @@
+import styles from "./showAll.module.css";
+
+const ShowAll = () => {
+  return <div>HELLO</div>;
+};
+
+export default ShowAll;
