+++ frontend/src/pages/selectContainer/selectRoute/selectRoute.jsx
@@ -1,0 +1,7 @@
+import styles from "./selectRoute.module.css";
+
+const SelectRoute = () => {
+  return <div></div>;
+};
+
+export default SelectRoute;
